Shared : move Cryptography to common
