+++ src/common/Cryptography/BigNumber.cpp
@@ -15,7 +15,7 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "Auth/BigNumber.h"
+#include "BigNumber.h"
 #include <openssl/bn.h>
 #include <algorithm>
