+++ src/common/Cryptography/Hmac.cpp
@@ -15,7 +15,7 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "Auth/Hmac.h"
+#include "Hmac.h"
 #include "BigNumber.h"
 
 #if defined(OPENSSL_VERSION_NUMBER) && OPENSSL_VERSION_NUMBER < 0x10100000L
