+++ src/server/game/CMakeLists.txt
@@ -41,6 +41,10 @@ include_directories(
   ${CMAKE_SOURCE_DIR}/dep/recastnavigation/Recast
   ${CMAKE_SOURCE_DIR}/src/server/shared
   ${CMAKE_SOURCE_DIR}/src/server/shared/Database
+  
+  ${CMAKE_SOURCE_DIR}/src/common/Cryptography/
+  ${CMAKE_SOURCE_DIR}/src/common/Cryptography/Authentication
+  
   ${CMAKE_CURRENT_SOURCE_DIR}
   
   ${CMAKE_CURRENT_SOURCE_DIR}/Accounts
