+++ src/server/game/Handlers/CharacterHandler.cpp
@@ -21,7 +21,7 @@
 #include "World.h"
 #include "WorldPacket.h"
 #include "WorldSession.h"
-#include "Auth/md5.h"
+#include "md5.h"
 #include "Database/DatabaseEnv.h"
 #include "Database/DatabaseImpl.h"
