+++ src/server/game/Server/WorldSocket.cpp
@@ -36,7 +36,7 @@
 #include "AddonHandler.h"
 #include "Opcodes.h"
 #include "Database/DatabaseEnv.h"
-#include "Auth/Sha1.h"
+#include "Sha1.h"
 #include "WorldSession.h"
 #include "WorldSocketMgr.h"
 #include "Log.h"
