+++ src/server/game/Server/WorldSocket.h
@@ -33,7 +33,7 @@
 #endif /* ACE_LACKS_PRAGMA_ONCE */
 
 #include "Common.h"
-#include "Auth/AuthCrypt.h"
+#include "AuthCrypt.h"
 
 class ACE_Message_Block;
 class WorldPacket;
