+++ src/server/game/Warden/WardenWin.cpp
@@ -15,8 +15,8 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "Auth/Hmac.h"
-#include "Auth/WardenKeyGeneration.h"
+#include "Hmac.h"
+#include "WardenKeyGeneration.h"
 #include "Common.h"
 #include "WorldPacket.h"
 #include "WorldSession.h"
